InfluxDB: Fix variable interpolation on adhoc filters
Grafana frontend code sends adhoc filters to the backend in both the
`tags` and `adhocFilters` params.

The values in `tags` have dashboard variables interpolated, while those
in `adhocFilters` don't. This PR switches backend handling to ignore the
`adhocFilters` param for adhoc filter handling, in favor of using
`tags`, so that variables are interpolated.